refactor(empty): Moved event row to a separate component so that use client can be moved down
